% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{changementPhase}
\alias{changementPhase}
\title{Extrait les le lignes où il y a passage de la phase expiratoire à
inspiratoire ou inversement}
\usage{
changementPhase(dataset)
}
\arguments{
\item{dataset}{Dataset returned by \code{servou::load_servou_data()}}
}
\description{
Extrait les le lignes où il y a passage de la phase expiratoire à
inspiratoire ou inversement
}
