% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cycle}
\alias{cycle}
\title{Extract one ventilation cycle, composet of one inspiration and one
expiration}
\usage{
cycle(d, i)
}
\arguments{
\item{d}{Dataset to extract the cycle from}

\item{i}{Rank in the sequence of the cycle to extrack}
}
\description{
Extract one ventilation cycle, composet of one inspiration and one
expiration
}
