% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumary.R
\name{cycles_preview}
\alias{cycles_preview}
\title{Display every ventilation cycle contained in
\code{filename}}
\usage{
cycles_preview(file)
}
\arguments{
\item{file}{Filename containing the data to analyse and display}
}
\description{
Display every ventilation cycle contained in
\code{filename}
}
