% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cyclesWithExpPause}
\alias{cyclesWithExpPause}
\title{Return the ventilation cycle containing an expiratory pause}
\usage{
cyclesWithExpPause(file)
}
\arguments{
\item{file}{File to parse}
}
\description{
Return the ventilation cycle containing an expiratory pause
}
