% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{cyclesWithInspPause}
\alias{cyclesWithInspPause}
\title{Return the ventilation cycle containing an inspiratory pause}
\usage{
cyclesWithInspPause(file)
}
\arguments{
\item{file}{File to parse}
}
\description{
Return the ventilation cycle containing an inspiratory pause
}
