% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dy.R
\name{dy}
\alias{dy}
\title{Evidence a variation (delta) in the y axis on a previously plotted
grahp}
\usage{
dy(x, y0, y1, lab)
}
\arguments{
\item{x}{Position in the x axis of the annotation}

\item{y0}{Starting value in the y axis}

\item{y1}{End value in the y axis}

\item{lab}{label to be printed beside the annotation}
}
\description{
Evidence a variation (delta) in the y axis on a previously plotted
grahp
}
