% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDecel.R
\name{expi}
\alias{expi}
\title{Shift a vector by 1 to the right, repeating the first value and
poppint the last value to keet the same length}
\usage{
expi(c)
}
\description{
Shift a vector by 1 to the right, repeating the first value and
poppint the last value to keet the same length
}
