% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumary.R
\name{files_table}
\alias{files_table}
\title{Create a table sumarizing all the Servo-U data file contained in a
folder}
\usage{
files_table(df)
}
\arguments{
\item{df}{Data frame to use for creation of the table or folder
path to create it}
}
\description{
Create a table sumarizing all the Servo-U data file contained in a
folder
}
