% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumary.R
\name{filessumary}
\alias{filessumary}
\title{Create a data frame sumarizing all the Servo-U data file contained
in a folder}
\usage{
filessumary(dossier)
}
\arguments{
\item{dossier}{Dossier contemamt les fichiers}
}
\description{
Create a data frame sumarizing all the Servo-U data file contained
in a folder
}
