% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPlot.R
\name{fit_decel_preview}
\alias{fit_decel_preview}
\title{Apply \code{servou::fitDecel()} to every ventilation cycle contained in
\code{filename} ant display a sheet of graphical representation of the
results}
\usage{
fit_decel_preview(file)
}
\arguments{
\item{file}{Filename containing the data to analyse and display}
}
\description{
Apply \code{servou::fitDecel()} to every ventilation cycle contained in
\code{filename} ant display a sheet of graphical representation of the
results
}
