% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDecel.R
\name{fitDecel}
\alias{fitDecel}
\title{Fit an inverse pababolic function to the inspiratory flow waveform
to calculate the \emph{flow index}.}
\usage{
fitDecel(cycle, aTreshold = 0.01, silent = TRUE)
}
\arguments{
\item{cycle}{Cycle to fit}

\item{aTreshold}{Treshold to pass to pp()}

\item{silent}{Wether the errors and warnings of snls should be supressed}
}
\description{
Fit an inverse pababolic function to the inspiratory flow waveform
to calculate the \emph{flow index}.
}
