% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPlot.R
\name{fitPlot}
\alias{fitPlot}
\title{Adjust an inverse exponential curve to the decelerating portion of
a flow waveform to compute the \emph{Flow Index}.}
\usage{
fitPlot(c, details = FALSE, title = NULL, posI = "graphic")
}
\arguments{
\item{c}{Ventilation cycle (one inspiration and one expiration)
returned by servou::cycle or servou::cycles.}
}
\description{
Adjust an inverse exponential curve to the decelerating portion of
a flow waveform to compute the \emph{Flow Index}.
}
