% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumary.R
\name{ftrig}
\alias{ftrig}
\title{Calculate the frequency of ventilator triggering by the patient}
\usage{
ftrig(file)
}
\arguments{
\item{file}{Path of the Servo-U data file to analyse}
}
\description{
Calculate the frequency of ventilator triggering by the patient
}
