% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_servou_meta.R
\name{load_servou_meta}
\alias{load_servou_meta}
\title{Load Servo-U ventilation parameters}
\usage{
load_servou_meta(filename)
}
\arguments{
\item{filename}{File name (including path) containing the data to
load. Must be a standard file exported by a Servo-U ventilator in
\emph{french}.}
}
\value{
t Data table of ventiation parameters.
}
\description{
Load Servo-U ventilation parameters
}
