% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_servou_recr_data.R
\name{load_servou_recr_data}
\alias{load_servou_recr_data}
\title{Load Servo-U data}
\usage{
load_servou_recr_data(filename)
}
\arguments{
\item{filename}{File name (including path) containing the data to load. Must
be a standard file exported by a Servo-U ventilator in \emph{french}.}
}
\value{
t Data table with the folowing columns: Durée, Phase, PRESSION,
DÉBIT, VOLUME, Trigger. Aditional columns may also be present
(e.g. EADI, Poes).
}
\description{
Load Servo-U data
}
