% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumary.R
\name{mean_fi}
\alias{mean_fi}
\title{Calculate the mean Flow index of ventilation cycles in a Servo-U
datafile}
\usage{
mean_fi(file)
}
\arguments{
\item{file}{Path of the Servo-U data file to analyse}
}
\description{
Calculate the mean Flow index of ventilation cycles in a Servo-U
datafile
}
