% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_settings.R
\name{parse_settings}
\alias{parse_settings}
\title{Parse Servo-U settings}
\usage{
parse_settings(filename)
}
\arguments{
\item{filename}{File name (including path) containing the data to
load. Must be a standard file exported by a Servo-U ventilator in
\emph{french}.}
}
\value{
t List of ventiator settings.
}
\description{
Parse Servo-U settings
}
