% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PmusIndex.R
\name{PMI}
\alias{PMI}
\title{Calculate the Pmus Index from an inspiratory pause}
\usage{
PMI(fichier, numPause = 1, debug = FALSE, silent = FALSE)
}
\arguments{
\item{fichier}{Filename of the Servo-U reccording file containing
an inspiratory pause}

\item{x}{Position of the annotation on the x axis}
}
\description{
Calculate the Pmus Index from an inspiratory pause
}
