% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PmusIndex.R
\name{PMIgraph}
\alias{PMIgraph}
\title{Draw a graph showing the calculation of the Pmus Index from an
inspiratory pause}
\usage{
PMIgraph(
  fichier,
  numPause = 1,
  x = NULL,
  offsetStart = 2.5,
  offsetEnd = 1,
  ai_as_title = FALSE,
  t = NULL,
  debug = FALSE
)
}
\arguments{
\item{fichier}{Filename of the Servo-U reccording file containing
an inspiratory pause}

\item{x}{Position of the annotation on the x axis}
}
\description{
Draw a graph showing the calculation of the Pmus Index from an
inspiratory pause
}
