% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PmusIndex.R
\name{PMIsheet}
\alias{PMIsheet}
\title{Draw a sheet of PMIgraph of every files in a folder}
\usage{
PMIsheet(path)
}
\arguments{
\item{path}{Folder containing the Servo-U data files}
}
\description{
Draw a sheet of PMIgraph of every files in a folder
}
