% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmusGraph.R
\name{pmusGraph}
\alias{pmusGraph}
\title{Create a graph evidencing the evaluating of Pmus by an expiratory
occlusion maneuver}
\usage{
pmusGraph(fichier, numPause = 1, x = NULL, ...)
}
\arguments{
\item{fichier}{Servou-U reccording file containing an expiratory
occlusion}

\item{numPause}{ventilationcycle inthe file with the expiratory
occlusion}

\item{x}{Where to put the arrow represingting Delta Pocc (in x
unit)}
}
\description{
Create a graph evidencing the evaluating of Pmus by an expiratory
occlusion maneuver
}
