% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmusGraph.R
\name{Pocc}
\alias{Pocc}
\title{Caculate the variation of airway pressure generation by a pateint
inspiratory effort during an expiratory occlusion maneuver (Δ P occ)}
\usage{
Pocc(fichier, numPause = 1)
}
\arguments{
\item{file}{Servou data file to analyse}
}
\description{
Caculate the variation of airway pressure generation by a pateint
inspiratory effort during an expiratory occlusion maneuver (Δ P occ)
}
