% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDecel.R
\name{pp}
\alias{pp}
\title{Select the part of the inspiratory flow waveform folowing the
inspiratory rise time.}
\usage{
pp(dataset, aTreshold = 0.01)
}
\description{
Select the part of the inspiratory flow waveform folowing the
inspiratory rise time.
}
