% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changementPhase.R
\name{cycles}
\alias{cycles}
\title{Return a list containing the dataset splited in ventilation
cycles}
\usage{
cycles(d)
}
\arguments{
\item{d}{Dataset returned by \code{servou::load_servou_data()}}
}
\value{
A list containing a dataset for each ventilation cycle
contained in the input dataset
}
\description{
Return a list containing the dataset splited in ventilation
cycles
}
