% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDecel.R
\name{fitDecel}
\alias{fitDecel}
\title{Fit an inverse pababolic function to the inspiratory flow waveform
to calculate the \emph{flow index}.}
\usage{
fitDecel(dataset)
}
\description{
Fit an inverse pababolic function to the inspiratory flow waveform
to calculate the \emph{flow index}.
}
